Upgrade more dependencies to their latest version.

+++ p2p/source/jsonrpc.hpp
@@ -96,7 +96,7 @@ struct Coded<Type_, typename std::enable_if<std::is_signed<Type_>::value>::type>
     }
 };
 
-template <unsigned Bits_, boost::multiprecision::cpp_int_check_type Check_>
+template <size_t Bits_, boost::multiprecision::cpp_int_check_type Check_>
 struct Coded<boost::multiprecision::number<boost::multiprecision::backends::cpp_int_backend<Bits_, Bits_, boost::multiprecision::unsigned_magnitude, Check_, void>>, typename std::enable_if<Bits_ % 8 == 0>::type> :
     public Numeric<false, (Bits_ >> 3), boost::multiprecision::number<boost::multiprecision::backends::cpp_int_backend<Bits_, Bits_, boost::multiprecision::unsigned_magnitude, Check_, void>>>
 {
@@ -105,7 +105,7 @@ struct Coded<boost::multiprecision::number<boost::multiprecision::backends::cpp_
     }
 };
 
-template <unsigned Bits_, boost::multiprecision::cpp_int_check_type Check_>
+template <size_t Bits_, boost::multiprecision::cpp_int_check_type Check_>
 struct Coded<boost::multiprecision::number<boost::multiprecision::backends::cpp_int_backend<Bits_, Bits_, boost::multiprecision::signed_magnitude, Check_, void>>, typename std::enable_if<Bits_ % 8 == 0>::type> :
     public Numeric<true, (Bits_ >> 3), boost::multiprecision::number<boost::multiprecision::backends::cpp_int_backend<Bits_, Bits_, boost::multiprecision::signed_magnitude, Check_, void>>>
 {
